Refactor to support integration with hunger app
We'll use a proxy/redirect on the webserver to swap handle hunger.json
requests with the actual hunger webapp rather than the static
mock-choice.json file. That means the mock file should be named the same
thing as the rewritten path for it to work both in local file mode,
static server mode, or the actual implementation without having to
change the json locations.