Use CORS instead of JSONP
I'm trying to get this working without really understanding what I'm doing, but the more I read, the more I get that jsonp is not the same thing as cors. The dropwizard Hunger app isn't setup to support jsonp at the moment, and I'm testing this against changes that have added cors headers, so that may be why it isn't working when I've included the jsonp option.