Upgrade to angular site
The actual site is hosted on google cloud/firebase hosting now, as is a
staging site -- this is kind of the staging site for the staging site.
We may end up turning it into a blog or data explorer in the future, but
for now we're just mirroring the staging site and retiring what once
was.